# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BlindCash API
  version: 1.0.0
  description: |
    BlindCash — 基于 Chaum 盲签名的 eCash 系统 API。

    **协议概述**（M2-M7）：
    - 取款走 4-move cut-and-choose 盲签名协议：init → submit → reveal → cancel
    - 支付走 token 兑付：POST /api/payment
    - 银行公钥可公开获取：GET /api/bank/pubkey

    **认证**：所有受保护接口需 `Authorization: Bearer <JWT>`。
    JWT 由 /api/auth/register 或 /api/auth/login 返回。

    **限流**（Phase 4）：
    - 全局：100 req/min/IP
    - 登录/注册：5 req/min/IP（防爆破）
    - 取款/支付：10 req/min/user

    **文献参考**：
    - [1] D. Chaum, "Blind Signatures for Untraceable Payments", CRYPTO 1982
    - [2] BIP-340 — Schnorr 签名（secp256k1）
    - [3] NIST SP 800-57rev5 §8.3.4 — 密钥轮换 cryptoperiod
servers:
  - url: http://localhost:4100
    description: 本地开发
  - url: /
    description: 相对路径（生产同源）
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security: []
paths:
...
  /api/withdraw/init:
    post:
      summary: ① 取款 init
      security:
        - bearerAuth: []
      description: |
        开启取款会话。限流：10 req/min/user（txLimiter）。
        4-move 协议第①步：扣余额 + 生成 N 个 k_i 承诺 R_i。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: integer
                  minimum: 1
      responses:
        '201':
          description: 会话创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  R:
                    type: array
                    items:
                      type: string
                  amount:
                    type: integer
                  N:
                    type: integer
                  ttl_ms:
                    type: integer
        '400':
          description: 金额无效或余额不足
        '409':
          description: 已有进行中的取款会话
  /api/withdraw/submit:
    post:
      summary: ③ 取款 submit
      security:
        - bearerAuth: []
      description: |
        提交 N 个盲化候选，银行随机选 j。
        α/β 绝不能出现在 payload 中（API defense）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, candidates]
              properties:
                session_id:
                  type: string
                candidates:
                  type: array
                  items:
                    type: object
                    properties:
                      e:
                        type: string
                      R_prime:
                        type: string
                      serial:
                        type: string
      responses:
        '200':
          description: 银行选中 j
          content:
            application/json:
              schema:
                type: object
                properties:
                  j:
                    type: integer
        '400':
          description: payload 含 α/β 或校验失败
  /api/withdraw/reveal:
    post:
      summary: ⑤ 取款 reveal + sign
      security:
        - bearerAuth: []
      description: 用户揭示 i≠j 的盲化因子，银行校验后签名第 j 个候选。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, revealed]
              properties:
                session_id:
                  type: string
                revealed:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: 返回签名 s_j
          content:
            application/json:
              schema:
                type: object
                properties:
                  s_j:
                    type: string
                  key_id:
                    type: integer
        '400':
          description: 揭示校验失败（作弊）
        '409':
          description: 会话已提交过
  /api/withdraw/cancel:
    post:
      summary: ⑦ 取款 cancel
      security:
        - bearerAuth: []
      description: 取消取款会话，退还余额。committed 状态不可取消。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties:
                session_id:
                  type: string
      responses:
        '200':
          description: 取消成功，已退款
        '400':
          description: 会话已 committed 不可取消
...